support 1.21.2-1.21.4/show llama strength before taming show health in decimal before taming fix wrong llama strength color

+++ src/client/java/monkey/lumpy/horse/stats/vanilla/gui/Tooltip.java
@@ -13,7 +13,7 @@ import net.minecraft.text.Text;
 public class Tooltip extends LightweightGuiDescription {
     private ModConfig config;
 
-    public Tooltip(double speed, double jump, int health) {
+    public Tooltip(double speed, double jump, double health) {
         super();
         config = AutoConfig.getConfigHolder(ModConfig.class).getConfig();
         WBox root = new WBox(Axis.VERTICAL);
